modules needed by wms should not use config_loader
