+++ mslib/mswms/mpl_hsec.py
@@ -42,7 +42,7 @@ import numpy as np
 import PIL.Image
 
 from mslib.mswms import mss_2D_sections
-from mslib.utils.coordinate import get_projection_params
+from mslib.utils.get_projection_params import get_projection_params
 from mslib.utils.units import convert_to
 from mslib.mswms.utils import make_cbar_labels_readable
 from mslib.utils.loggerdef import configure_mpl_logger
